Move KeyValueElement to common
